2023/04/10

+++ Lemo/src/main/java/kr/co/Lemo/dao/MyDAO.java
@@ -47,21 +47,21 @@ public interface MyDAO {
 
     // @since 2023/03/28
     public int selectTotalReviews(SearchCondition sc);
-    public ReviewVO selectReview(@Param("res_no") int res_no, @Param("user_id") String user_id);
-    public ReviewVO selectReviewAccommodation(@Param("res_no") int res_no, @Param("user_id") String user_id);
+    public ReviewVO selectReview(@Param("res_no") long res_no, @Param("user_id") String user_id);
+    public ReviewVO selectReviewAccommodation(@Param("res_no") long res_no, @Param("user_id") String user_id);
     public int insertReview(Map<String, Object> param);
 
     // @since 2023/03/29
-    public ReservationVO selectReservation(@Param("res_no") int res_no, @Param("user_id") String user_id);
+    public ReservationVO selectReservation(@Param("res_no") long res_no, @Param("user_id") String user_id);
     public int deleteReservation(long res_no);
-    public int selectCheckReview(int res_no);
-    public int selectCheckDiary(int res_no);
+    public int selectCheckReview(long res_no);
+    public int selectCheckDiary(long res_no);
 
     // @since 2023/03/30
     public int updateReservationState(long res_no);
 
     // @since 2023/03/31
-    public ProductAccommodationVO selectDiaryXY(int res_no);
+    public ProductAccommodationVO selectDiaryXY(long res_no);
     public String selectReservationImpUid(long res_no);
 
     // @since 2023/04/05
